requirements.txt
flask
flask-socketio
python-socketio
redis
google-generativeai
transformers
torch
gitpython
flask-login